+++ crypto/stark/src/constraints/mod.rs
@@ -1,4 +1,3 @@
 pub mod boundary;
 pub mod evaluator;
-pub mod lookup;
 pub mod transition;
