+++ crypto/stark/src/examples/mod.rs
@@ -1,11 +1,9 @@
-pub mod add_with_lookup;
 pub mod bit_flags;
-pub mod cpu_with_lookup;
 pub mod dummy_air;
 pub mod fibonacci_2_cols_shifted;
 pub mod fibonacci_2_columns;
 pub mod fibonacci_rap;
-pub mod mul_with_lookup;
+pub mod multi_table_lookup;
 pub mod quadratic_air;
 pub mod read_only_memory;
 pub mod read_only_memory_logup;
