--- crypto/stark/src/examples/mul_with_lookup.rs
@@ -1,46 +1,0 @@
-use crate::{
-    constraints::{
-        lookup::{
-            AirWithLookup, AuxiliaryTraceBuildData, NullBoundaryConstraintBuilder, TableInteraction,
-        },
-        transition::TransitionConstraint,
-    },
-    proof::options::ProofOptions,
-    trace::TraceTable,
-};
-use math::field::fields::fft_friendly::{
-    babybear::Babybear31PrimeField, quartic_babybear::Degree4BabyBearExtensionField,
-};
-type F = Babybear31PrimeField;
-type E = Degree4BabyBearExtensionField;
-
-pub fn new_mul_air_with_lookup(
-    trace: &TraceTable<F, E>,
-    proof_options: &ProofOptions,
-) -> AirWithLookup<F, E, NullBoundaryConstraintBuilder, ()> {
-    // TODO: define mul-specific constraints here
-    let transition_constraints: Vec<Box<dyn TransitionConstraint<F, E>>> = vec![];
-
-    let step_size = 1;
-
-    let auxiliary_trace_build_data = AuxiliaryTraceBuildData {
-        interactions: vec![
-            // Interaction with CPU table
-            TableInteraction {
-                // multiplicity column
-                flag_columns: vec![3],
-                // values a , b, c
-                value_columns: vec![0, 1, 2],
-            },
-        ],
-    };
-
-    AirWithLookup::create(
-        trace,
-        auxiliary_trace_build_data,
-        proof_options,
-        step_size,
-        transition_constraints,
-        (),
-    )
-}
